Moved pipeline-selections related methods out of GoConfigService to PipelineSelectionsService
