+++ server/src/com/thoughtworks/go/server/service/GoConfigService.java
@@ -30,8 +30,8 @@ import com.thoughtworks.go.config.validation.GoConfigValidity;
 import com.thoughtworks.go.domain.*;
 import com.thoughtworks.go.domain.config.Admin;
 import com.thoughtworks.go.domain.materials.MaterialConfig;
-import com.thoughtworks.go.domain.packagerepository.PackageRepositories;
 import com.thoughtworks.go.domain.packagerepository.PackageDefinition;
+import com.thoughtworks.go.domain.packagerepository.PackageRepositories;
 import com.thoughtworks.go.domain.packagerepository.PackageRepository;
 import com.thoughtworks.go.domain.scm.SCM;
 import com.thoughtworks.go.i18n.LocalizedMessage;
@@ -42,7 +42,6 @@ import com.thoughtworks.go.presentation.TriStateSelection;
 import com.thoughtworks.go.server.cache.GoCache;
 import com.thoughtworks.go.server.domain.PipelineConfigDependencyGraph;
 import com.thoughtworks.go.server.domain.Username;
-import com.thoughtworks.go.server.domain.user.PipelineSelections;
 import com.thoughtworks.go.server.initializers.Initializer;
 import com.thoughtworks.go.server.persistence.PipelineRepository;
 import com.thoughtworks.go.server.security.GoAcl;
@@ -806,58 +805,6 @@ public class GoConfigService implements Initializer, CruiseConfigProvider {
         return new PipelineConfigDependencyGraph(currentPipeline, graphs.toArray(new PipelineConfigDependencyGraph[0]));
     }
 
-    public PipelineSelections getSelectedPipelines(String id, Long userId) {
-        PipelineSelections pipelineSelections = getPersistedPipelineSelections(id, userId);
-        if (pipelineSelections == null) {
-            pipelineSelections = PipelineSelections.ALL;
-        }
-        return pipelineSelections;
-    }
-
-    public long persistSelectedPipelines(String id, Long userId, List<String> selectedPipelines, boolean isBlacklist) {
-        PipelineSelections pipelineSelections = findOrCreateCurrentPipelineSelectionsFor(id, userId);
-
-        if (isBlacklist) {
-            List<String> unselectedPipelines = invertSelections(selectedPipelines);
-            pipelineSelections.update(unselectedPipelines, clock.currentTime(), userId, isBlacklist);
-        } else {
-            pipelineSelections.update(selectedPipelines, clock.currentTime(), userId, isBlacklist);
-        }
-
-        return pipelineRepository.saveSelectedPipelines(pipelineSelections);
-    }
-
-    private PipelineSelections findOrCreateCurrentPipelineSelectionsFor(String id, Long userId) {
-        PipelineSelections pipelineSelections = isSecurityEnabled() ? pipelineRepository.findPipelineSelectionsByUserId(userId) : pipelineRepository.findPipelineSelectionsById(id);
-        if (pipelineSelections == null) {
-            pipelineSelections = new PipelineSelections(new ArrayList<>(), clock.currentTime(), userId, true);
-        }
-        return pipelineSelections;
-    }
-
-    private List<String> invertSelections(List<String> selectedPipelines) {
-        List<String> unselectedPipelines = new ArrayList<>();
-        List<PipelineConfig> pipelineConfigList = cruiseConfig().getAllPipelineConfigs();
-        for (PipelineConfig pipelineConfig : pipelineConfigList) {
-            String pipelineName = CaseInsensitiveString.str(pipelineConfig.name());
-            if (!selectedPipelines.contains(pipelineName)) {
-                unselectedPipelines.add(pipelineName);
-            }
-        }
-        return unselectedPipelines;
-    }
-
-    private PipelineSelections getPersistedPipelineSelections(String id, Long userId) {
-        PipelineSelections pipelineSelections = null;
-        if (isSecurityEnabled()) {
-            pipelineSelections = pipelineRepository.findPipelineSelectionsByUserId(userId);
-        }
-        if (pipelineSelections == null) {
-            pipelineSelections = pipelineRepository.findPipelineSelectionsById(id);
-        }
-        return pipelineSelections;
-    }
-
     public List<Role> rolesForUser(final CaseInsensitiveString user) {
         return security().getRoles().memberRoles(new AdminUser(user));
     }
@@ -981,12 +928,12 @@ public class GoConfigService implements Initializer, CruiseConfigProvider {
         return null;
     }
 
-    public void updateUserPipelineSelections(String id, Long userId, CaseInsensitiveString pipelineToAdd) {
-        PipelineSelections currentSelections = findOrCreateCurrentPipelineSelectionsFor(id, userId);
-        if (!currentSelections.isBlacklist()) {
-            currentSelections.addPipelineToSelections(pipelineToAdd);
-            pipelineRepository.saveSelectedPipelines(currentSelections);
-        }
+    public boolean isAuthorizedToEditTemplate(String templateName, Username username) {
+        return isUserAdmin(username) || getCurrentConfig().getTemplates().canUserEditTemplate(templateName, username.getUsername());
+    }
+
+    public boolean isAuthorizedToViewAndEditTemplates(Username username) {
+        return getCurrentConfig().getTemplates().canViewAndEditTemplate(username.getUsername());
     }
 
     public boolean isPipelineEditable(String pipelineName) {
